long overdue hotfixes

+++ server/controllers/commentControllers.js
@@ -2,10 +2,12 @@ const Comment = require("../db/models/Comment");
 const User = require("../db/models/User");
 
 exports.createComment = async (req, res) => {
-  const { programId, userId, organizationId, body, time } = req.body;
+  const { programId, body, time } = req.body;
+  const userId = req.session.userId || null;
+  const organizationId = req.session.organizationId || null;
   // console.log(programId,userId,body,time);
 
-  if (!programId || !body || !time) {
+  if (!programId || !body || !time || (!userId && !organizationId)) {
     return res.sendStatus(400);
   }
 
@@ -28,6 +30,12 @@ exports.showComment = async (req, res) => {
 exports.updateComment = async (req, res) => {
   const { body } = req.body;
   const {id} = req.params;
+  const existingComment = await Comment.findById(id);
+  if (!existingComment) return res.sendStatus(404);
+  const ownsComment = (existingComment.user_id && Number(existingComment.user_id) === Number(req.session.userId))
+    || (existingComment.organization_id && Number(existingComment.organization_id) === Number(req.session.organizationId));
+  if (!ownsComment) return res.sendStatus(403);
+
   const comment = await Comment.update(id, body);
   res.send(comment);
 };
@@ -40,6 +48,12 @@ exports.getAllCommentsOfUser = async (req, res) => {
 
 exports.deleteComment = async (req,res) => {
   const {id} = req.params;
+  const existingComment = await Comment.findById(id);
+  if (!existingComment) return res.sendStatus(404);
+  const ownsComment = (existingComment.user_id && Number(existingComment.user_id) === Number(req.session.userId))
+    || (existingComment.organization_id && Number(existingComment.organization_id) === Number(req.session.organizationId));
+  if (!ownsComment) return res.sendStatus(403);
+
   const comment = await Comment.deleteComment(id);
   res.send(comment);
 };
